pipeline/pyproject.toml
[project]
name = "pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "coremltools>=9.0",
    "fastapi>=0.129.2",
    "ffmpeg-python>=0.2.0",
    "firebase-admin>=7.1.0",
    "imagehash>=4.3.2",
    "onnxruntime>=1.24.2",
    "pillow>=12.1.1",
    "ultralytics>=8.4.14",
    "uvicorn>=0.41.0",
]
